Move each function to its separate file
