+++ tsconfig.json
@@ -6,6 +6,6 @@
     "outDir": "./dist",
     "strict": true
   },
-  "include": ["./lib/**/*"],
-  "exclude": ["./lib/**/*.spec.ts"]
+  "include": ["src/*"],
+  "exclude": ["node_modules", "tests/*.spec.ts"]
 }
